requirements.txt
fastapi==0.115.8
python-dotenv==1.0.1
supabase==2.13.0
uvicorn==0.34.0
pydantic-to-typescript==2.0.0


